builtinの結合
