+++ srcs/utils/env_util.c
@@ -1,0 +1,30 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   env_util.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yumiyao <yumiyao@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/05/22 14:05:05 by yumiyao           #+#    #+#             */
+/*   Updated: 2025/05/22 14:06:15 by yumiyao          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	is_valid_env(char *name)
+{
+	int	i;
+
+	i = 0;
+	if (!ft_isalpha(name[0]) && name[0] != '_')
+		return (0);
+	while (name[i])
+	{
+		if (ft_isalnum(name[i]) || name[i] == '_')
+			++i;
+		else
+			return (0);
+	}
+	return (1);
+}
